pyproject.toml
[project]
name = "restock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "mcp>=1.21.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.38.0",
]
[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true

dependencies = []
